+++ src/game/Entity.cpp
@@ -99,7 +99,7 @@ Entity::Entity(const res::path & classPath, EntityInstance instance)
 	obj = NULL;
 	std::fill_n(anims, MAX_ANIMS, (ANIM_HANDLE *)NULL);
 	
-	for(long l = 0; l < MAX_ANIM_LAYERS; l++) {
+	for(size_t l = 0; l < MAX_ANIM_LAYERS; l++) {
 		animlayer[l] = AnimLayer();
 	}
